refactor tests: centralize login helper
