+++ tests/test_cat_photo.py
@@ -1,23 +1,12 @@
 import re
+from conftest import login_helper
 
 
 def test_cat_photo_display_after_login(client):
     username = "catuser"
     password = "secret"
 
-    client.post(
-        "/signup",
-        data={"username": username, "password": password},
-        follow_redirects=False,
-    )
-
-    response = client.post(
-        "/login",
-        data={"username": username, "password": password},
-        follow_redirects=False,
-    )
-    assert response.status_code == 303
-    client.cookies.update(response.cookies)
+    login_helper(client, username, password)
 
     response = client.get("/protected")
     assert response.status_code == 200
