+++ tests/test_darkmode.py
@@ -1,3 +1,6 @@
+from conftest import login_helper
+
+
 def test_darkmode_toggle_not_in_login_page(client):
     response = client.get("/login")
     assert response.status_code == 200
@@ -17,20 +20,8 @@ def test_darkmode_toggle_in_account_page_after_login(client):
     username = "darktoggle"
     password = "secret"
 
-    # create user if not exists
-    client.post(
-        "/signup",
-        data={"username": username, "password": password},
-        follow_redirects=False,
-    )
-
-    response = client.post(
-        "/login",
-        data={"username": username, "password": password},
-        follow_redirects=False,
-    )
-    assert response.status_code == 303
-    client.cookies.update(response.cookies)
+    # create user if not exists and log in
+    login_helper(client, username, password)
 
     response = client.get("/account")
     assert response.status_code == 200
